+++ projects/solid-forms-react/src/public-api.ts
@@ -1,4 +1,3 @@
-export * from './withControl';
 export * from './useControl';
 export * from './useControlState';
 export * from 'solid-forms/core';